細々としてところ修正

+++ PlayList.cpp
@@ -556,7 +556,7 @@ extern COggDlg *og;
 extern CString filen,fnn;
 
 extern int modesub,ret2;
-extern __int64 loop1, loop2;
+extern int loop1, loop2;
 
 void CPlayList::Get(int i)
 {
